Refactor rasterizer internal API

+++ include/ttf/rasterizer.hpp
@@ -27,7 +27,8 @@ class rasterizer
     rasterizer & operator=(rasterizer const &) = delete;
     rasterizer & operator=(rasterizer &&);
 
-    void rasterize(shape const & s, int x, int y) const;
+    void rasterize(shape const & s, float x, float y) const;
+    // void rasterize(shape const & s, int x, int y) const;
 
     private:
     class implementation;
